[#41] [BE] Rename : 'psql.js' to 'db.js'
