+++ backend/src/repositories/articleRepository.js
@@ -1,4 +1,4 @@
-import pool from '../config/psql.js';
+import pool from '../config/db.js';
 
 /**
  * article_id, title, subtitle, content, author_id, created_at, updated_at, is_public
